+++ src/pages/Swap/Swap.tsx
@@ -20,6 +20,7 @@ import { useHasPriceData } from '../../lib/tokenPrices';
 import { getRouterEstimates, useRouterResult } from './hooks/useRouter';
 import { useSwap } from './hooks/useSwap';
 
+import { formatAmount } from '../../lib/utils/number';
 import { getAmountInDenom } from '../../lib/web3/utils/tokens';
 import { formatLongPrice } from '../../lib/utils/number';
 import { cleanInput } from '../../components/TokenInputGroup/utils';
@@ -86,11 +87,13 @@ export default function Swap() {
   const onFormSubmit = useCallback(
     function (event?: React.FormEvent<HTMLFormElement>) {
       if (event) event.preventDefault();
-      if (address && routerResult && tokenA && tokenB) {
+      const tolerance = parseFloat(slippage) / 100;
+      if (address && routerResult && tokenA && tokenB && !isNaN(tolerance)) {
         // convert to swap request format
         const result = routerResult;
         // Cosmos requires tokens in integer format of smallest denomination
-        // todo: add slippage tolerance setting into API request
+        // add slippage tolerance
+        const minOut = result.amountOut.multipliedBy(1 - tolerance);
         swapRequest({
           amountIn:
             getAmountInDenom(
@@ -101,11 +104,10 @@ export default function Swap() {
             ) || '0',
           tokenIn: result.tokenIn,
           tokenOut: result.tokenOut,
-          // TODO: add tolerance factor
           minOut:
             getAmountInDenom(
               tokenB,
-              result.amountOut,
+              minOut,
               tokenB?.display,
               tokenB?.display
             ) || '0',
@@ -113,7 +115,7 @@ export default function Swap() {
         });
       }
     },
-    [address, routerResult, tokenA, tokenB, swapRequest]
+    [address, routerResult, tokenA, tokenB, slippage, swapRequest]
   );
 
   const onValueAChanged = useCallback((newValue: string) => {
@@ -172,6 +174,17 @@ export default function Swap() {
   }, [tokenA, tokenB]);
 
   const hasPriceData = useHasPriceData([tokenA, tokenB]);
+
+  const priceImpact =
+    routerResult &&
+    routerResult.priceIn?.isGreaterThan(0) &&
+    routerResult.priceOut?.isGreaterThan(0)
+      ? new BigNumber(100).minus(
+          new BigNumber(routerResult.priceOut)
+            .dividedBy(new BigNumber(routerResult.priceIn))
+            .multipliedBy(100)
+        )
+      : undefined;
 
   const tradeCard = (
     <div className="trade-card">
@@ -305,7 +318,31 @@ export default function Swap() {
                   )}
                 </span>
                 <span className="text-header">Price Impact</span>
-                <span className="text-value">0.00%</span>
+                {priceImpact && (
+                  <span
+                    className={[
+                      'text-value',
+                      (() => {
+                        switch (true) {
+                          case priceImpact.isGreaterThanOrEqualTo(0):
+                            return 'text-success';
+                          case priceImpact.isGreaterThan(-1):
+                            return 'text-value';
+                          case priceImpact.isGreaterThan(-5):
+                            return 'text';
+                          default:
+                            return 'text-error';
+                        }
+                      })(),
+                    ].join(' ')}
+                  >
+                    {formatAmount(priceImpact.toFixed(), {
+                      maximumSignificantDigits: 4,
+                      minimumSignificantDigits: 4,
+                    })}
+                    %
+                  </span>
+                )}
               </div>
             )}
         </div>